restructure code into a valid python package
snake case package name
