+++ super_duper_octo_barnacle/tests/test_main.py
@@ -1,7 +1,6 @@
-from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
-from main import app
+from super_duper_octo_barnacle.main import app
 
 client = TestClient(app)
